zeno_agent/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
authlib==1.6.4
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
cmdstanpy==1.2.5
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
docstring-parser==0.17.0
fastapi==0.116.2
fonttools==4.60.0
google-adk==1.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.41.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.115.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.38.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
holidays==0.81
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
matplotlib==3.10.6
mcp==1.14.1
narwhals==2.6.0
numpy==2.3.3
nvidia-nccl-cu12==2.28.3
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pandas==2.3.2
patsy==1.0.1
pillow==11.3.0
plotly==6.3.0
prophet==1.1.7
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
sse-starlette==3.0.2
stanio==0.5.1
starlette==0.48.0
statsmodels==0.14.5
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.36.0
websockets==15.0.1
xgboost==3.0.5
zipp==3.23.0
